Delete errors midd and fix controllers

+++ backend/src/controllers/ClientController.js
@@ -1,6 +1,4 @@
 import { Client } from '../models/Client.js';
-import BusinessError from '../middlewares/BusinessError.js';
-
 
 export const getClient = async ( req, res ) =>{
     try {
@@ -15,12 +13,14 @@ export const getOneClient = async ( req, res ) =>{
     try {
         const { id } = req.params;
         if(!id){
-            throw new BusinessError({message:'El id es obligatorio', code: 400});
+            return res.status( 400 ).json({ message: "El id es obligatorio" }); 
         }
         const getOneClient = await Client.findByPk( id );
+
         if(!getOneClient){
-            throw new BusinessError({message:'El cliente no existe', code: 404});
+            return res.status( 404 ).json({ message: "El cliente no existe" }); 
         }
+
         return res.status( 200 ).json({ result: getOneClient }); 
     } catch ( error ) {
         return res.status( 500 ).json({ message: error }); 
@@ -31,28 +31,36 @@ export const createClient = async (req, res) =>{
     try {
         const { name, phone, address, email, pass, profile_picture } = req.body;
 
+        if(!name || !phone || !address || !email || !pass){
+            return res.status( 400 ).json({ message: "El cuerpo de la solicitud está incompleto. Debes proporcionar todos los parámetros requeridos" }); 
+        }
+
         const clientExist = await Client.findOne({ where: { email } });
 
         if(clientExist) {
-            console.log("entro")
-            throw new BusinessError({message:'El cliente ya existe', code: 400});
+            return res.status( 400 ).json({ message: "El cliente ya existe"})
         }
         
         const createdClient = await Client.create({ name, phone, address, email, pass, profile_picture });
 
         return res.status( 201 ).json({ result: createdClient });
     } catch ( error ) {
-        return res.status( 500 ).json({ message: error.message });
+        return res.status( 500 ).json({ message: error });
     }
 }
 
 export const updateClient = async (req, res) =>{
     try {
         const { id } = req.params;
+
+        if(!id){
+            return res.status( 400 ).json({ message: "El id es obligatorio" }); 
+        }
+
         const clientToUpdate = await Client.findByPk( id );
 
         if(!clientToUpdate){
-            throw new BusinessError({message:'El cliente no existe', code: 404});
+            return res.status( 404 ).json({ message: 'El cliente no existe' });
         }
 
         clientToUpdate.set( req.body );
@@ -66,6 +74,11 @@ export const updateClient = async (req, res) =>{
 export const deleteClient = async (req, res) =>{
     try {
         const { id } = req.params;
+
+        if(!id){
+            return res.status( 400 ).json({ message: "El id es obligatorio" }); 
+        }
+        
         const deleteClient = await Client.destroy({
             where: {
                 id
