+++ backend/src/controllers/CompanyController.js
@@ -1,24 +1,23 @@
 import { Company } from '../models/Company.js';
-import BusinessError from '../middlewares/BusinessError.js';
 
 export const getCompany = async (req ,res) =>{
     try {
         const getCompany = await Company.findAll();
         return res.status( 200 ).json({ result: getCompany }); 
     } catch ( error ) {
-         next();
+        return res.status( 500 ).json({ message: error });
     }
 }
 export const getOneCompany = async (req ,res) =>{
     try {
         const { id } = req.params;
         if( !id ) {
-            throw new BusinessError({message:'El id es obligatorio', code: 400});
+            return res.status( 400 ).json({ message: 'El id es obligatorio' });
         }
         const getOneCompany = await Company.findByPk( id );
 
         if( !getOneCompany ) {
-            throw new BusinessError({message:'La compañia no existe', code: 404});
+            return res.status( 404 ).json({ message: 'La compañia no existe' });
         }
         return res.status( 200 ).json({ result: getOneCompany }); 
     } catch ( error ) {
@@ -29,24 +28,39 @@ export const createCompany = async (req ,res) =>{
     try {
         const { name } = req.body;
 
-        const companyExist = await Company.findOne({ where: { name } });
-        if(companyExist) {
-            throw new BusinessError({message:'La compañia ya existe', code: 400});
+        if(!name){
+            return res.status( 400 ).json({ message: "El cuerpo de la solicitud está incompleto. Debes proporcionar todos los parámetros requeridos" }); 
         }
 
-        const createCompany = await Company.create({ name });
-        return res.status( 200 ).json({ result: createCompany });
+        const companyExist = await Company.findOne({ where: { name } });
+        if(companyExist) {
+            return res.status( 400 ).json({ message: 'La compañia ya existe' });
+        }
+
+        const createdCompany = await Company.create({ name });
+        return res.status( 200 ).json({ result: createdCompany });
     } catch ( error ) {
-        return res.status( 500 ).json({ message: error.message });
+        return res.status( 500 ).json({ message: error });
     }
 }
 export const updateCompany = async (req ,res) =>{
     try {
         const { id } = req.params;
-        const updateCompany = await Company.findByPk( id );
-        updateCompany.set( req.body );
-        updateCompany.save();
-        return res.status( 200 ).json({ result: updateCompany });
+        
+        if(!id){
+            return res.status( 400 ).json({ message: 'El id es obligatorio' });
+        }
+
+        const companyToUpdate = await Company.findByPk( id );
+
+        if(!companyToUpdate){
+            return res.status( 404 ).json({ message: 'La empresa no existe' });
+        }
+
+        companyToUpdate.set( req.body );
+
+        companyToUpdate.save();
+        return res.status( 200 ).json({ result: companyToUpdate });
     } catch ( error ) {
         return res.status( 500 ).json({ message: error });
     }
@@ -54,11 +68,17 @@ export const updateCompany = async (req ,res) =>{
 export const deleteCompany = async (req ,res) =>{
     try {
         const { id } = req.params;
+
+        if(!id){
+            return res.status( 400 ).json({ message: 'El id es obligatorio' });
+        }
+
         const deleteCompany = await Company.destroy({
             where: {
                 id
             }
         });
+        
         return res.status( 200 ).json({ result: deleteCompany })
     } catch ( error ) {
         return res.status( 500 ).json({ message: error });
