+++ backend/src/controllers/ServiceController.js
@@ -13,7 +13,13 @@ export const getServices = async ( req, res ) =>{
 export const getService = async (req, res) => {
     try {
         const { id } = req.params;
+
+        if(!id){
+            return res.status( 400 ).json({ message: "El id es obligatorio" }); 
+        }
+        
         const getService = await Service.findByPk( id );
+
         return res.status( 200 ).json({ result: getService }); 
     } catch (error) {
         return res.status( 500 ).json({ message: error });
@@ -27,12 +33,12 @@ export const createService = async (req, res) => {
         const serviceExist = await Service.findOne( {where: {name}} )
 
         if(serviceExist){
-            throw new BusinessError({message:'El servicio ya existe', code: 400});
+            return res.status( 400 ).json({ message: 'El servicio ya existe' });
         }
         
-        const cretedService = await Service.create({name, description})
+        const createdService = await Service.create({name, description})
 
-        return res.status( 201 ).json({ result: createService });
+        return res.status( 201 ).json({ result: createdService });
     } catch (error) {
         return res.status( 500 ).json({ message: error });
     }
@@ -42,14 +48,18 @@ export const createService = async (req, res) => {
 export const updateService = async(req, res) => {
     try {
         const { id } = req.params
-        const { name, description } = req.body
+
+        if(!id){
+            return res.status( 400 ).json({ message: "El id es obligatorio" }); 
+        }
+        
 
         const serviceToUpdate = await Service.findByPk( id );
         if(!serviceToUpdate){
             throw new BusinessError({message:'El servicio no existe', code: 404});
         }
 
-        serviceToUpdate.set( {name, description} );
+        serviceToUpdate.set( req.body );
         serviceToUpdate.save();
 
         return res.status( 200 ).json({ result: serviceToUpdate });
@@ -63,6 +73,9 @@ export const deleteService = async(req, res) => {
     try{
         const { id } = req.params;
 
+        if(!id){
+            return res.status( 400 ).json({ message: "El id es obligatorio" }); 
+        }
         const deletedService = await Service.destroy({
             where: {
                 id
